rename: ディレクトリの誤字を修正
PrefChecboxes => PrefCheckboxes
