+++ src/components/elements/PageViewer/PageViewer.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { usePrefectures } from "@/hooks/usePrefectures";
-import { PrefCheckboxes } from "../PrefChecboxes/PrefCheckboxes";
+import { PrefCheckboxes } from "../PrefCheckboxes/PrefCheckboxes";
 import { PopulationResponse } from "@/types/populations/populations";
 import { PrefecturesResponse } from "@/types/prefectures/prefectures";
 import PopulationLine from "../PopulationLine/PopulationLine";
